% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_templates.R
\name{create_initial_discussion_prompt}
\alias{create_initial_discussion_prompt}
\title{Create prompt for the initial round of discussion}
\usage{
create_initial_discussion_prompt(
  cluster_id,
  cluster_genes,
  tissue_name,
  initial_predictions
)
}
\description{
Create prompt for the initial round of discussion
}
\keyword{internal}
