% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_model_manager.R
\docType{data}
\name{custom_providers}
\alias{custom_providers}
\title{Custom model manager for mLLMCelltype}
\format{
An object of class \code{environment} of length 0.
}
\usage{
custom_providers
}
\description{
This module provides functionality to register and manage custom LLM
providers
and models. It allows users to integrate their own LLM services with the
mLLMCelltype framework.
}
\keyword{Environment}
\keyword{and}
\keyword{configurations}
\keyword{custom}
\keyword{internal}
\keyword{providers}
\keyword{store}
\keyword{their}
\keyword{to}
