% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facilitate_cluster_discussion.R
\name{facilitate_cluster_discussion}
\alias{facilitate_cluster_discussion}
\title{Facilitate discussion for a controversial cluster}
\usage{
facilitate_cluster_discussion(
  cluster_id,
  input,
  tissue_name,
  models,
  api_keys,
  initial_predictions,
  top_gene_count,
  max_rounds = 3,
  controversy_threshold = 0.7,
  entropy_threshold = 1,
  consensus_check_model = NULL,
  base_urls = NULL
)
}
\description{
Facilitate discussion for a controversial cluster
}
\note{
This function uses create_initial_discussion_prompt and create_discussion_prompt from prompt_templates.R
}
\keyword{internal}
