% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facilitate_cluster_discussion.R
\name{filter_valid_responses}
\alias{filter_valid_responses}
\title{Filter out error responses from model round responses}
\usage{
filter_valid_responses(responses, cluster_id, round = NULL)
}
\description{
Filter out error responses from model round responses
}
\keyword{internal}
