% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_consensus.R
\name{find_majority_prediction}
\alias{find_majority_prediction}
\title{Find majority prediction from response lines}
\usage{
find_majority_prediction(lines)
}
\description{
Find majority prediction from response lines
}
\keyword{internal}
