% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grok_processor.R
\name{GrokProcessor}
\alias{GrokProcessor}
\title{Grok API Processor}
\description{
Grok API Processor

Grok API Processor
}
\details{
Concrete implementation of BaseAPIProcessor for Grok models.
Handles Grok-specific API calls, authentication, and response parsing.
}
\section{Super class}{
\code{mLLMCelltype::BaseAPIProcessor} -> \code{GrokProcessor}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_url}}{Grok API endpoint URL}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GrokProcessor-new}{\code{GrokProcessor$new()}}
\item \href{#method-GrokProcessor-make_api_call}{\code{GrokProcessor$make_api_call()}}
\item \href{#method-GrokProcessor-extract_response_content}{\code{GrokProcessor$extract_response_content()}}
\item \href{#method-GrokProcessor-clone}{\code{GrokProcessor$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mLLMCelltype" data-topic="BaseAPIProcessor" data-id="process_request"><a href='../../mLLMCelltype/html/BaseAPIProcessor.html#method-BaseAPIProcessor-process_request'><code>mLLMCelltype::BaseAPIProcessor$process_request()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GrokProcessor-new"></a>}}
\if{latex}{\out{\hypertarget{method-GrokProcessor-new}{}}}
\subsection{Method \code{new()}}{
Initialize Grok processor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GrokProcessor$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GrokProcessor-make_api_call"></a>}}
\if{latex}{\out{\hypertarget{method-GrokProcessor-make_api_call}{}}}
\subsection{Method \code{make_api_call()}}{
Make API call to Grok
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GrokProcessor$make_api_call(chunk_content, model, api_key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chunk_content}}{Content for this chunk}

\item{\code{model}}{Model identifier}

\item{\code{api_key}}{API key}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
httr response object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GrokProcessor-extract_response_content"></a>}}
\if{latex}{\out{\hypertarget{method-GrokProcessor-extract_response_content}{}}}
\subsection{Method \code{extract_response_content()}}{
Extract response content from Grok API response
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GrokProcessor$extract_response_content(response, model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{response}}{httr response object}

\item{\code{model}}{Model identifier}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Extracted text content
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GrokProcessor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GrokProcessor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GrokProcessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
