% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_annotation.R
\name{identify_controversial_clusters}
\alias{identify_controversial_clusters}
\title{Identify controversial clusters based on consensus analysis}
\usage{
identify_controversial_clusters(
  input,
  individual_predictions,
  controversy_threshold,
  entropy_threshold,
  api_keys,
  consensus_check_model = NULL,
  base_urls = NULL
)
}
\description{
Identify controversial clusters based on consensus analysis
}
\keyword{internal}
