% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_model_manager.R
\name{list_custom_models}
\alias{list_custom_models}
\title{Get list of registered custom models}
\usage{
list_custom_models()
}
\value{
Character vector of model names
}
\description{
Get list of registered custom models
}
