% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_consensus.R
\name{normalize_annotation}
\alias{normalize_annotation}
\title{Normalize annotation for comparison}
\usage{
normalize_annotation(annotation)
}
\description{
Normalize annotation for comparison
}
\keyword{internal}
