% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_consensus.R
\name{parse_flexible_format}
\alias{parse_flexible_format}
\title{Parse flexible format consensus response}
\usage{
parse_flexible_format(lines)
}
\description{
Parse flexible format consensus response
}
\keyword{internal}
