% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_consensus.R
\name{prepare_models_list}
\alias{prepare_models_list}
\title{Prepare list of models to try for consensus checking}
\usage{
prepare_models_list(consensus_check_model = NULL)
}
\arguments{
\item{consensus_check_model}{User-specified model (can be NULL)}
}
\value{
Character vector of models in order of preference
}
\description{
Prepare list of models to try for consensus checking
}
\keyword{internal}
