% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_annotation.R
\name{process_controversial_clusters}
\alias{process_controversial_clusters}
\title{Process controversial clusters through discussion}
\usage{
process_controversial_clusters(
  controversial_clusters,
  input,
  tissue_name,
  successful_models,
  api_keys,
  individual_predictions,
  top_gene_count,
  controversy_threshold,
  entropy_threshold,
  max_discussion_rounds,
  cache_manager,
  use_cache,
  consensus_check_model = NULL,
  force_rerun = FALSE,
  base_urls = NULL
)
}
\description{
Process controversial clusters through discussion
}
\keyword{internal}
