% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_zhipu.R
\name{process_zhipu}
\alias{process_zhipu}
\title{Process request using Zhipu models}
\usage{
process_zhipu(prompt, model, api_key, base_url = NULL)
}
\description{
Process request using Zhipu models
}
\keyword{internal}
