% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_utils.R
\name{sanitize_base_url}
\alias{sanitize_base_url}
\title{Sanitize base URL}
\usage{
sanitize_base_url(url)
}
\arguments{
\item{url}{URL to sanitize}
}
\value{
Sanitized URL
}
\description{
Sanitize base URL
}
\keyword{internal}
