% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepfun_processor.R
\name{StepFunProcessor}
\alias{StepFunProcessor}
\title{StepFun API Processor}
\description{
StepFun API Processor

StepFun API Processor
}
\details{
Concrete implementation of BaseAPIProcessor for StepFun models.
Handles StepFun-specific API calls, authentication, and response parsing.
}
\section{Super class}{
\code{mLLMCelltype::BaseAPIProcessor} -> \code{StepFunProcessor}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_url}}{StepFun API endpoint URL}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StepFunProcessor-new}{\code{StepFunProcessor$new()}}
\item \href{#method-StepFunProcessor-make_api_call}{\code{StepFunProcessor$make_api_call()}}
\item \href{#method-StepFunProcessor-extract_response_content}{\code{StepFunProcessor$extract_response_content()}}
\item \href{#method-StepFunProcessor-clone}{\code{StepFunProcessor$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mLLMCelltype" data-topic="BaseAPIProcessor" data-id="process_request"><a href='../../mLLMCelltype/html/BaseAPIProcessor.html#method-BaseAPIProcessor-process_request'><code>mLLMCelltype::BaseAPIProcessor$process_request()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StepFunProcessor-new"></a>}}
\if{latex}{\out{\hypertarget{method-StepFunProcessor-new}{}}}
\subsection{Method \code{new()}}{
Initialize StepFun processor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StepFunProcessor$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StepFunProcessor-make_api_call"></a>}}
\if{latex}{\out{\hypertarget{method-StepFunProcessor-make_api_call}{}}}
\subsection{Method \code{make_api_call()}}{
Make API call to StepFun
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StepFunProcessor$make_api_call(chunk_content, model, api_key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chunk_content}}{Content for this chunk}

\item{\code{model}}{Model identifier}

\item{\code{api_key}}{API key}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
httr response object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StepFunProcessor-extract_response_content"></a>}}
\if{latex}{\out{\hypertarget{method-StepFunProcessor-extract_response_content}{}}}
\subsection{Method \code{extract_response_content()}}{
Extract response content from StepFun API response
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StepFunProcessor$extract_response_content(response, model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{response}}{httr response object}

\item{\code{model}}{Model identifier}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Extracted text content
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StepFunProcessor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StepFunProcessor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StepFunProcessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
