% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_annotation.R
\name{clean_annotation}
\alias{clean_annotation}
\title{Clean annotation text by removing prefixes and extra whitespace}
\usage{
clean_annotation(annotation)
}
\arguments{
\item{annotation}{The annotation text to clean}
}
\value{
Cleaned annotation text
}
\description{
Clean annotation text by removing prefixes and extra whitespace
}
\keyword{internal}
