% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_annotation.R
\name{combine_results}
\alias{combine_results}
\title{Combine results from all phases of consensus annotation}
\usage{
combine_results(initial_results, controversy_results, discussion_results)
}
\arguments{
\item{initial_results}{Results from initial prediction phase}

\item{controversy_results}{Results from controversy identification phase}

\item{discussion_results}{Results from discussion phase}
}
\value{
Combined results
}
\description{
Combine results from all phases of consensus annotation
}
\keyword{internal}
