% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unified_logger.R
\name{configure_logger}
\alias{configure_logger}
\title{Set global logger configuration}
\usage{
configure_logger(level = "INFO", console_output = TRUE, json_format = TRUE)
}
\arguments{
\item{level}{Logging level}

\item{console_output}{Whether to output to console}

\item{json_format}{Whether to use JSON format}
}
\description{
Set global logger configuration
}
