% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_templates.R
\name{create_initial_discussion_prompt}
\alias{create_initial_discussion_prompt}
\title{Create prompt for the initial round of discussion}
\usage{
create_initial_discussion_prompt(
  cluster_id,
  cluster_genes,
  tissue_name,
  initial_predictions
)
}
\arguments{
\item{cluster_id}{The ID of the cluster being analyzed}

\item{cluster_genes}{The marker genes for the cluster}

\item{tissue_name}{The name of the tissue (optional)}

\item{initial_predictions}{A list of initial model predictions}
}
\value{
A formatted prompt string for the initial discussion round
}
\description{
Create prompt for the initial round of discussion
}
\keyword{internal}
