% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_annotation.R
\name{get_initial_predictions}
\alias{get_initial_predictions}
\title{Get initial predictions from all models}
\usage{
get_initial_predictions(
  input,
  tissue_name,
  models,
  api_keys,
  top_gene_count,
  base_urls = NULL
)
}
\description{
This function retrieves initial cell type predictions from all specified models.
It is an internal helper function used by the interactive_consensus_annotation function.
}
\keyword{internal}
