% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_annotation.R
\name{identify_controversial_clusters}
\alias{identify_controversial_clusters}
\title{Identify controversial clusters based on consensus analysis}
\usage{
identify_controversial_clusters(
  input,
  individual_predictions,
  controversy_threshold,
  entropy_threshold,
  api_keys,
  consensus_check_model = NULL
)
}
\arguments{
\item{input}{Either the differential gene table or a list of genes}

\item{individual_predictions}{List of predictions from each model}

\item{controversy_threshold}{Threshold for marking clusters as controversial}

\item{entropy_threshold}{Entropy threshold for identifying controversial clusters}
}
\value{
A list containing controversial clusters and consensus results
}
\description{
Identify controversial clusters based on consensus analysis
}
\keyword{internal}
