% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_annotation.R
\docType{package}
\name{mLLMCelltype-package}
\alias{mLLMCelltype}
\alias{mLLMCelltype-package}
\title{mLLMCelltype: Cell Type Annotation Using Large Language Models}
\description{
Automated cell type annotation for single-cell RNA sequencing data using consensus predictions from multiple large language models. Integrates with Seurat objects and provides uncertainty quantification for annotations. Supports various LLM providers including OpenAI, Anthropic, and Google. For details see Yang et al. (2025) \doi{10.1101/2025.04.10.647852}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cafferyang.com/mLLMCelltype/}
  \item Report bugs at \url{https://github.com/cafferychen777/mLLMCelltype/issues}
}

}
\author{
\strong{Maintainer}: Chen Yang \email{cafferychen777@tamu.edu} [copyright holder]

}
\keyword{internal}
