% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_consensus.R
\name{parse_consensus_response}
\alias{parse_consensus_response}
\title{Parse consensus response from model}
\usage{
parse_consensus_response(response)
}
\description{
Parse consensus response from model
}
\keyword{internal}
