% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_consensus.R
\name{parse_standard_format}
\alias{parse_standard_format}
\title{Parse standard 4-line consensus response format}
\usage{
parse_standard_format(result_lines)
}
\description{
Parse standard 4-line consensus response format
}
\keyword{internal}
