% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_consensus.R
\name{prepare_models_list}
\alias{prepare_models_list}
\title{Prepare list of models to try for consensus checking}
\usage{
prepare_models_list(consensus_check_model = NULL)
}
\description{
Prepare list of models to try for consensus checking
}
\keyword{internal}
