% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_anthropic.R
\name{process_anthropic}
\alias{process_anthropic}
\title{Process request using Anthropic models}
\usage{
process_anthropic(prompt, model, api_key, base_url = NULL)
}
\description{
Process request using Anthropic models
}
\keyword{internal}
