% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_deepseek.R
\name{process_deepseek}
\alias{process_deepseek}
\title{Process request using DeepSeek models}
\usage{
process_deepseek(prompt, model, api_key, base_url = NULL)
}
\description{
Process request using DeepSeek models
}
\keyword{internal}
