% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_openai.R
\name{process_openai}
\alias{process_openai}
\title{Process request using OpenAI models}
\usage{
process_openai(prompt, model, api_key, base_url = NULL)
}
\description{
Process request using OpenAI models
}
\keyword{internal}
