% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_stepfun.R
\name{process_stepfun}
\alias{process_stepfun}
\title{Process request using Stepfun models}
\usage{
process_stepfun(prompt, model, api_key)
}
\description{
Process request using Stepfun models
}
\keyword{internal}
