% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_annotation.R
\name{select_best_prediction}
\alias{select_best_prediction}
\title{Select the best prediction from consensus results}
\usage{
select_best_prediction(consensus_result, valid_predictions)
}
\arguments{
\item{consensus_result}{Consensus analysis result}

\item{valid_predictions}{Valid predictions for the cluster}
}
\value{
The best prediction
}
\description{
Select the best prediction from consensus results
}
\keyword{internal}
